+++ tests/Js/Main/TypeTests.fs
@@ -351,6 +351,53 @@ type InfoB = {
     Bar: string
 }
 
+[<Mangle>]
+type IUpperMangledInterface =
+    abstract Upper: string -> string
+    abstract Ping: unit -> string
+    abstract Value: string with get, set
+    abstract ReadOnlyValue: string with get
+    abstract SetterOnlyValue: string with set
+
+type UpperMangledClass() =
+    let mutable innerValue = ""
+
+    interface IUpperMangledInterface with
+        member _.Upper s = s.ToUpper()
+        member _.Ping() = "pong"
+
+        member _.Value
+            with get () = innerValue
+            and set v = innerValue <- v
+
+        member _.ReadOnlyValue = innerValue
+        member _.SetterOnlyValue
+            with set v = innerValue <- v
+
+[<Mangle>]
+type IGenericMangledInterface<'T> =
+    abstract Upper: string -> string
+    abstract Ping: unit -> string
+    abstract Value: 'T with get, set
+    abstract ReadOnlyValue: string with get
+    abstract SetterOnlyValue: 'T with set
+
+type GenericMangledClass() =
+    let mutable innerValue = ""
+
+    interface IGenericMangledInterface<string> with
+        member _.Upper s = s.ToUpper()
+        member _.Ping() = "pong"
+
+        member _.Value
+            with get () = innerValue
+            and set v = innerValue <- v
+
+        member _.ReadOnlyValue = innerValue
+
+        member _.SetterOnlyValue
+            with set v = innerValue <- v
+
 #if !FABLE_COMPILER_TYPESCRIPT
 [<AbstractClass>]
 type InfoAClass(info: InfoA) =
@@ -370,7 +417,7 @@ type FooInterface =
     abstract Item: int -> char with get, set
     abstract Sum: [<ParamArray>] items: string[] -> string
 
-[<Fable.Core.Mangle>]
+[<Mangle>]
 type BarInterface =
     abstract Bar: string with get, set
     abstract DoSomething: f: (float -> float -> float) * v: float -> float
@@ -1322,4 +1369,28 @@ let tests =
         let mutable arr = [| 1; 2; 3 |]
         let result = genericByrefFunc &arr
         result |> equal 3
+
+    testCase "mangled method on interface works"
+    <| fun () ->
+        let upper = UpperMangledClass()
+        (upper :> IUpperMangledInterface).Upper("hello") |> equal "HELLO"
+        (upper :> IUpperMangledInterface).Ping() |> equal "pong"
+        (upper :> IUpperMangledInterface).Value |> equal ""
+        (upper :> IUpperMangledInterface).Value <- "value"
+        (upper :> IUpperMangledInterface).Value |> equal "value"
+        (upper :> IUpperMangledInterface).ReadOnlyValue |> equal "value"
+        (upper :> IUpperMangledInterface).SetterOnlyValue <- "setter only value"
+        (upper :> IUpperMangledInterface).Value |> equal "setter only value"
+
+    testCase "mangled method on generic interface works"
+    <| fun () ->
+        let upper = GenericMangledClass()
+        (upper :> IGenericMangledInterface<string>).Upper("hello") |> equal "HELLO"
+        (upper :> IGenericMangledInterface<string>).Ping() |> equal "pong"
+        (upper :> IGenericMangledInterface<string>).Value |> equal ""
+        (upper :> IGenericMangledInterface<string>).Value <- "value"
+        (upper :> IGenericMangledInterface<string>).Value |> equal "value"
+        (upper :> IGenericMangledInterface<string>).ReadOnlyValue |> equal "value"
+        (upper :> IGenericMangledInterface<string>).SetterOnlyValue <- "setter only value"
+        (upper :> IGenericMangledInterface<string>).Value |> equal "setter only value"
   ]